change type to int

+++ src/main/java/phonerate/Gold.java
@@ -2,14 +2,14 @@ package phonerate;
 
 public class Gold {
 	private double goldrateperadditionalminute = 0.45;
-	private double goldincludedminutes = 1000;
+	private int goldincludedminutes = 1000;
 	private double goldrateperadditionalline = 14.50;
 	private double goldplanrate = 49.95;
 	
 	public double rateperadditionalminute() {
 		return goldrateperadditionalminute;
 	}
-	public double includedminutes() {
+	public int includedminutes() {
 		return goldincludedminutes;
 	}
 	public double rateperadditionalline() {
